requirements.txt
sentence-transformers==2.2.2
langchain
langchain-core
langchain-pinecone
langchain-community
langchain-huggingface
flask
pypdf
python-dotenv
pinecone[grpc]
requests
pydantic
gunicorn
-e .